refactor: ディレクトリ分割を見直し
- features/を`prefectures`と`compositionChart`に分割
