+++ src/features/compositionChart/components/ui/SelectableLabels.module.scss
