+++ src/features/compositionChart/constants/index.ts
@@ -6,8 +6,5 @@ export const CHART_LABELS = [
   '老年人口'
 ] as const
 
-/** 都道府県情報のクエリパラーメーターのキー名 */
-export const PREFECTURE_QUERY_KEY = 'prefecture'
-
 /** 人口構成の種類のクエリパラーメーターのキー名 */
 export const LABEL_QUERY_KEY = 'label'
