+++ src/features/compositionChart/index.ts
@@ -1,0 +1,2 @@
+export * from './components/ChartContainer'
+export * from './components/SelectableLabels'
