--- src/features/prefectureChart/api/index.ts
@@ -1,2 +1,0 @@
-export * from './getCompositionPerYear'
-export * from './getPrefectures'
