--- src/features/prefectureChart/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useCompositionPerYear'
-export * from './usePrefectureQuery'
