--- src/features/prefectureChart/index.ts
@@ -1,4 +1,0 @@
-export * from './components/SelectablePrefecturesContainer'
-export * from './components/SelectableCompositionLabels'
-export * from './components/CompositionChartContainer'
-export * from './api/getPrefectures'
