+++ src/features/prefectures/constants/index.ts
@@ -1,0 +1,2 @@
+/** 都道府県情報のクエリパラーメーターのキー名 */
+export const PREFECTURE_QUERY_KEY = 'prefecture'
